ingestion/requirements.txt
snowflake-connector-python[pandas]
requests
beautifulsoup4
edgartools
python-dotenv
pandas
